puts initialization into one file and allows setting cluster ID env var manually
