--- integration/01_circle_sha.go
@@ -1,29 +1,0 @@
-// +build k8srequired
-
-package integration
-
-import (
-	"fmt"
-	"os"
-)
-
-const (
-	// EnvVarCircleSHA is the process environment variable representing the
-	// CIRCLE_SHA1 env var.
-	EnvVarCircleSHA = "CIRCLE_SHA1"
-)
-
-var (
-	circleSHA string
-)
-
-func init() {
-	circleSHA = os.Getenv(EnvVarCircleSHA)
-	if circleSHA == "" {
-		panic(fmt.Sprintf("env var '%s' must not be empty", EnvVarCircleSHA))
-	}
-}
-
-func CircleSHA() string {
-	return circleSHA
-}
